+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Added
+* Added "Remove All Expressions" and "Collapse All" buttons to Live Watch panel, similar to the standard Watch view.
+
 ## [1.14.1] – 2025‑12‑15
 
 ### Fixed